third_party/blink/web_tests/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-choose-default-value-after-set-value-expected.txt

Test if change event fires when the user selects the default value after the value was changed by JS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.value is "2001-01-01T01:01:01"
PASS eventsCounter.input is undefined.
PASS eventsCounter.change is undefined.
==> "input" event was dispatched.
==> "change" event was dispatched.
PASS input.value is "2000-01-01T01:01:01"
PASS eventsCounter.input is 1
PASS eventsCounter.change is 1
PASS successfullyParsed is true

TEST COMPLETE
